Smart_Code_Tutor/requirements.txt
fastapi
uvicorn[standard]
websockets
langchain
langchain-openai
langchain-community
langchain-text-splitters
e2b
python-multipart
pydantic
python-dotenv
aiofiles
faiss-cpu